fix some potential error

+++ app/models/user.rb
@@ -2,8 +2,7 @@ class User < ActiveRecord::Base
   has_one :profile
   after_create :create_profile
   before_destroy :destroy_profile
-  # Include default devise modules. Others available are:
-  # :confirmable, :lockable, :timeoutable and :omniauthable
+  
   devise :database_authenticatable,
             :registerable,
             :recoverable,
@@ -18,6 +17,6 @@ class User < ActiveRecord::Base
   end
 
   def destroy_profile
-    profile.destroy
+    profile.destroy if profile.present?
   end
 end
